backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0
mysqlclient==2.2.7
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
pytest==8.3.5
pytest-django==4.9.0
